Update Rating property to Float

+++ app/src/main/java/com/example/roompractice/data/entity/JuiceEntity.kt
@@ -10,5 +10,5 @@ data class JuiceEntity(
     val name: String,
     val description: String,
     val color: String,
-    val rating: Int
+    val rating: Float
 )
